requirements.txt
numpy
trimesh >= 3.9.0
cherrypy
pillow
scipy
transforms3d
torch